feat(tt): reduce duplication
