+++ apps/total-typescript/src/server/routers/exercise.ts
@@ -7,15 +7,11 @@ import {
   ExerciseMedia,
   ExerciseMediaSchema,
   ExerciseSchema,
-  getExercise,
   getExerciseMedia,
 } from '../../lib/exercises'
-import {getToken} from 'next-auth/jwt'
-import {getSubscriberFromCookie} from '../../utils/get-convertkit-subscriber-from-cookie'
-import {getCurrentAbility, UserSchema} from '../../ability/ability'
-import {getWorkshop} from '../../lib/workshops'
-import {getTutorial} from '../../lib/tutorials'
-import {getSection} from '../../lib/sections'
+
+import {createAppAbility} from '../../ability/ability'
+import {ContentRulesSchema, getContentRules} from '../get-content-rules'
 
 export const exerciseRouter = createRouter()
   .query('bySlug', {
@@ -31,47 +27,15 @@ export const exerciseRouter = createRouter()
     },
   })
   .query('media', {
-    input: z.object({
-      moduleSlug: z.string().optional(),
-      moduleType: z.string().optional(),
-      lessonSlug: z.string().optional(),
-      sectionSlug: z.string().optional(),
-      isSolution: z.boolean().optional(),
-      muxPlaybackId: z.nullable(z.string().optional()),
-    }),
+    input: ContentRulesSchema,
     async resolve({ctx, input}) {
-      const token = await getToken({req: ctx.req})
-      const convertkitSubscriber = getSubscriberFromCookie(ctx.req)
-      const {
-        moduleSlug,
-        moduleType,
-        lessonSlug,
-        sectionSlug,
-        isSolution = false,
-        muxPlaybackId,
-      } = input
-      const module = moduleSlug
-        ? moduleType === 'workshop'
-          ? await getWorkshop(moduleSlug)
-          : await getTutorial(moduleSlug)
-        : undefined
-      const lesson = lessonSlug ? await getExercise(lessonSlug) : undefined
-      const section = sectionSlug ? await getSection(sectionSlug) : undefined
+      const rules = await getContentRules(ctx.req, input)
+      const ability = createAppAbility(rules)
 
-      const ability = getCurrentAbility({
-        ...(token && {user: UserSchema.parse(token)}),
-        ...(convertkitSubscriber && {
-          subscriber: convertkitSubscriber,
-        }),
-        module,
-        lesson,
-        section,
-        isSolution,
-        muxPlaybackId,
-      })
-
-      if (lessonSlug && ability.can('view', 'Content')) {
-        const exerciseMedia: ExerciseMedia = await getExerciseMedia(lessonSlug)
+      if (input.lessonSlug && ability.can('view', 'Content')) {
+        const exerciseMedia: ExerciseMedia = await getExerciseMedia(
+          input.lessonSlug,
+        )
         return ExerciseMediaSchema.parse(exerciseMedia)
       }
     },
